ZZALOG 3.4.26 Bug: Decode of I2/UR5ZA incorrectly selected Ukraine. Fix: Decode of I2 only looked for prefix I2 not I. Re-parse if callsign changes as well as record pointer. Bug: Exceptions trying to access DxAtlas when not connected. Bug: Empty band/mode value added to list of buttons

+++ zzalog/src/qso_details.cpp
@@ -207,11 +207,11 @@ qso_details::table_q::table_q(int X, int Y, int W, int H, const char* L) :
 	row_header(false);
 	col_header(true);
 	cols(4);
-	float wd = tiw;
-	col_width(0, wd * 0.3 );
-	col_width(1, wd * 0.2 );
-	col_width(2, wd * 0.2 );
-	col_width(3, wd * 0.3);
+	float wd = (float)tiw;
+	col_width(0, (int)(wd * 0.3) );
+	col_width(1, (int)(wd * 0.2) );
+	col_width(2, (int)(wd * 0.2) );
+	col_width(3, (int)(wd * 0.3) );
 	end();
 }
